Reestruturando ui com dbClient
